requirements.txt
# ============================================================
# Capstone_AI2 requirements.txt
# ============================================================
 
# 기존
sentence-transformers==2.7.0
scikit-learn==1.6.1
pandas==2.2.2
numpy==1.26.4
seaborn==0.13.2
matplotlib==3.8.4
torch==2.3.0
joblib==1.4.2
 
# 새로 추가 (FastAPI + LLM API)
fastapi
uvicorn[standard]
openai
python-dotenv
pika
boto3
